+++ src/HyperTrack/data_types/Location.ts
@@ -1,0 +1,7 @@
+export type Location = {
+  type: 'location';
+  value: {
+    latitude: number;
+    longitude: number;
+  };
+};