+++ src/HyperTrack/data_types/internal/DeviceId.ts
@@ -1,0 +1,4 @@
+export type DeviceId = {
+  type: 'deviceID';
+  value: string;
+};