+++ src/HyperTrack/data_types/internal/DeviceName.ts
@@ -1,0 +1,4 @@
+export type DeviceName = {
+  type: 'deviceName';
+  value: string;
+};