+++ src/HyperTrack/data_types/internal/IsTracking.ts
@@ -1,0 +1,4 @@
+export type IsTracking = {
+  type: 'isTracking';
+  value: boolean;
+};