+++ src/index.tsx
@@ -2,10 +2,7 @@ import HyperTrack from './HyperTrack/HyperTrack';
 export default HyperTrack;
 export { HyperTrack };
 
-export {
-  Location,
-  Error,
-  SDKError,
-  GeotagError,
-  LocationError,
-} from './HyperTrack/HyperTrack';
+export { HyperTrackError } from './HyperTrack/data_types/HyperTrackError';
+export { LocationError, NotRunning, Starting, Errors } from './HyperTrack/data_types/LocationError';
+export { Location } from './HyperTrack/data_types/Location';
+export { SdkInitParams } from './HyperTrack/data_types/SdkInitParams'