+++ CHANGELOG.md
@@ -1,8 +1,30 @@
 # Changelog
-All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [9.0.0] - 2023-02-17
+
+### Changed
+- Updated HyperTrack iOS SDK to 4.14.0
+- Updated HyperTrack Android SDK to 6.4.0
+- `createInstance()` renamed to `initialize()`
+- `getDeviceID()` renamed to `getDeviceId()`
+- `syncDeviceSettings()` renamed to `sync()`
+- `setDeviceName()` renamed to `setName()`
+- `subscribeToErrors()` callback now has param type (`HyperTrackError`)
+- `LocationError` now has different structure
+
+### Added
+- `initialize()` configuration params for
+  - Debug logging
+  - Background location permissions request for Android
+  - Mock locations
+- `subscribeToAvailability()`
+
+### Removed
+- `enableDebugLogging()` (use `initialize()` param `loggingEnabled` istead)
+- `setTripMarker()` (use `addGeotag` instead)
 
 ## [8.2.2] - 2022-10-10
 ### Fixed